Hide Hashing and Caching MSDFs from user. some renamings and refactors packing unpacking of minUV+textureID
