+++ 62_CAD/CMakeLists.txt
@@ -11,8 +11,6 @@ set(EXAMPLE_SOURCES
   "${CMAKE_CURRENT_SOURCE_DIR}/Polyline.h"
   "${CMAKE_CURRENT_SOURCE_DIR}/DrawResourcesFiller.cpp"
   "${CMAKE_CURRENT_SOURCE_DIR}/DrawResourcesFiller.h"
-  "${CMAKE_CURRENT_SOURCE_DIR}/MSDFs.cpp"
-  "${CMAKE_CURRENT_SOURCE_DIR}/MSDFs.h"
   "../../src/nbl/ext/TextRendering/TextRendering.cpp" # TODO: this one will be a part of dedicated Nabla ext called "TextRendering" later on which uses MSDF + Freetype
 )
 set(EXAMPLE_INCLUDES
